# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenASM API",
        "description": "OpenASM (Open Asset Management) REST API for comprehensive asset management with multi-tenancy support",
        "contact": {
            "name": "OpenASM Support",
            "email": "admin@openasm.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "OpenASM API Server"
        }
    ],
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user and get JWT token",
                "operationId": "8b1ca12fe82a8104a6fbb70f5e149956",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "JWT access token",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "description": "Token expiration time in seconds",
                                            "type": "integer"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "roles": {
                                            "description": "User roles",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "permissions": {
                                            "description": "User permissions",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh JWT token",
                "operationId": "c42c3cd7fd7fe0dbd5bcc42962059e90",
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "New JWT access token",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "description": "Token expiration time in seconds",
                                            "type": "integer"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "roles": {
                                            "description": "User roles",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "permissions": {
                                            "description": "User permissions",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token refresh failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get authenticated user information",
                "operationId": "37d33f0281f67d31d05b4c45eea0143b",
                "responses": {
                    "200": {
                        "description": "User information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "roles": {
                                            "description": "User roles",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "permissions": {
                                            "description": "User permissions",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "organization": {
                                            "description": "User's organization",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user and invalidate token",
                "operationId": "f6b8957b4c0b25636613a9f0512a8ba9",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string"
                    },
                    "error": {
                        "description": "Detailed error information",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Organization": {
                "properties": {
                    "id": {
                        "description": "Organization unique identifier",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "Organization name",
                        "type": "string"
                    },
                    "code": {
                        "description": "Organization code",
                        "type": "string"
                    },
                    "email": {
                        "description": "Organization email",
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "phone": {
                        "description": "Organization phone",
                        "type": "string",
                        "nullable": true
                    },
                    "website": {
                        "description": "Organization website",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "description": {
                        "description": "Organization description",
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "description": "Organization address",
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "description": "Organization city",
                        "type": "string",
                        "nullable": true
                    },
                    "state": {
                        "description": "Organization state",
                        "type": "string",
                        "nullable": true
                    },
                    "country": {
                        "description": "Organization country",
                        "type": "string",
                        "nullable": true
                    },
                    "postal_code": {
                        "description": "Organization postal code",
                        "type": "string",
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Whether the organization is active",
                        "type": "boolean"
                    },
                    "metadata": {
                        "description": "Additional organization metadata",
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "first_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "last_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "department": {
                        "type": "string",
                        "nullable": true
                    },
                    "organization_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "organization": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Organization"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}